Resolved #1140 

Assembly became a little faster.

                | before    | after
------------- | ----------- | --------------
bundles    | 1875ms | **1820ms**
examples | 3414ms | **2968ms**